// src/app/models/GeneralMember.ts
export class GeneralMember{
  id?: number;
  personId?: number;
  parentId?: number;
  memberCode?: string;
  personName?: string;
  parentName?: string;
  remark?: string;
  areaDescription?: string;
  email?: string;
  personTypeName?: string;
  age?: number;
  gender?: string;
  mobile1?: string;
  mobile2?: string;
  voterId?: string;
  assemblyName?: string;
  pollingNumber?: number;
  totalWorkers?: number;
}